Fix some C-style casts an some types usage.

+++ td/utils/Heap.h
@@ -74,8 +74,8 @@ class KHeap {
   }
 
   void check() const {
-    for (int i = 0; i < (int)array_.size(); i++) {
-      for (int j = i * K + 1; j < i * K + 1 + K && j < (int)array_.size(); j++) {
+    for (size_t i = 0; i < array_.size(); i++) {
+      for (size_t j = i * K + 1; j < i * K + 1 + K && j < array_.size(); j++) {
         CHECK(array_[i].key_ <= array_[j].key_) << i << " " << j;
       }
     }
